Moved unique ID processing to a separate file
