+++ src/render.rs
@@ -1,67 +1,8 @@
 use std::borrow::Cow;
 use std::collections::HashMap;
 
+use crate::id::unique_id_from_content;
 use crate::{resolve::AdmonitionMeta, types::CssId};
-
-//--------------------------------------------
-// FIXME: Copied the code from mdbook as-is...!!
-
-use regex::Regex;
-use std::sync::LazyLock;
-
-/// Convert the given string to a valid HTML element ID.
-/// The only restriction is that the ID must not contain any ASCII whitespace.
-fn normalize_id(content: &str) -> String {
-    content
-        .chars()
-        .filter_map(|ch| {
-            if ch.is_alphanumeric() || ch == '_' || ch == '-' {
-                Some(ch.to_ascii_lowercase())
-            } else if ch.is_whitespace() {
-                Some('-')
-            } else {
-                None
-            }
-        })
-        .collect::<String>()
-}
-
-/// Generate an ID for use with anchors which is derived from a "normalised"
-/// string.
-fn id_from_content(content: &str) -> String {
-    let mut content = content.to_string();
-
-    // Skip any tags or html-encoded stuff
-    static HTML: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"(<.*?>)").unwrap());
-    content = HTML.replace_all(&content, "").into();
-    const REPL_SUB: &[&str] = &["&lt;", "&gt;", "&amp;", "&#39;", "&quot;"];
-    for sub in REPL_SUB {
-        content = content.replace(sub, "");
-    }
-
-    // Remove spaces and hashes indicating a header
-    let trimmed = content.trim().trim_start_matches('#').trim();
-    normalize_id(trimmed)
-}
-
-/// Generate an ID for use with anchors which is derived from a "normalised"
-/// string.
-///
-/// Each ID returned will be unique, if the same `id_counter` is provided on
-/// each call.
-pub fn unique_id_from_content(content: &str, id_counter: &mut HashMap<String, usize>) -> String {
-    let id = id_from_content(content);
-
-    // If we have headers with the same normalized id, append an incrementing counter
-    let id_count = id_counter.entry(id.clone()).or_insert(0);
-    let unique_id = match *id_count {
-        0 => id,
-        id_count => format!("{id}-{id_count}"),
-    };
-    *id_count += 1;
-    unique_id
-}
-//--------------------------------------------
 
 #[derive(Debug, PartialEq)]
 pub(crate) struct Admonition<'a> {
